Updated tooling.

+++ Build-Artifact.ps1
@@ -1,5 +1,9 @@
 $ArticlesToPublish = Get-Content .\posts\posts.json | ConvertFrom-Json
 
+if(($ArticlesToPublish | Where-Object { $_.posts.action -eq "publish"}).posts.Count -eq 0 ) {
+    throw "No articles to publish"
+}
+
 if(Test-Path -Path .\publish) { Remove-Item -Path .\publish -Recurse -Force }
 New-Item -Name publish -ItemType Directory
 New-Item -Path .\publish -Name "posts" -ItemType Directory
